De-duplicate EntityIntenseExplosion  (#480)
