+++ config/symbols.beta.txt
@@ -98,3 +98,5 @@ D_8003C854 = 0x8003C854;
 D_80072E9C = 0x80072E9C;
 D_80072EB4 = 0x80072EB4;
 D_80096ED0 = 0x80096ED0;
+
+g_InitializeEntityData0 = 0x80180508;
