+++ config/symbols.us.stcen.txt
@@ -1,5 +1,6 @@
 g_SpriteBanks = 0x80180040;
 PfnEntityUpdates = 0x80180394;
+g_InitializeEntityData0 = 0x8018041C;
 UNK_Invincibility0 = 0x801807B8;
 g_Rooms = 0x801812D4;
 EntityBackgroundBlock = 0x8018D600;
