+++ config/symbols.us.stdre.txt
@@ -1,6 +1,7 @@
 g_SpriteBanks = 0x80180040;
 PfnEntityUpdates = 0x801803C8;
 g_eBreakableInit = 0x80180458;
+g_InitializeEntityData0 = 0x80180470;
 g_eBreakableAnimations = 0x80180610;
 g_eBreakableHitboxes = 0x80180630;
 g_eBreakableExplosionTypes = 0x80180638;
