+++ config/symbols.us.stnp3.txt
@@ -2,6 +2,7 @@ g_SpriteBanks = 0x80180040;
 g_pStObjLayout = 0x80180728;
 PfnEntityUpdates = 0x801808D0;
 g_eBreakableInit = 0x80180A3C;
+g_InitializeEntityData0 = 0x80180A54;
 g_eBreakableAnimations = 0x80180E28;
 g_eBreakableHitboxes = 0x80180E48;
 g_eBreakableExplosionTypes = 0x80180E50;
