+++ config/symbols.us.strwrp.txt
@@ -1,5 +1,6 @@
 g_SpriteBanks = 0x80180040;
 PfnEntityUpdates = 0x801803E4;
+g_InitializeEntityData0 = 0x80180458;
 UNK_Invincibility0 = 0x80180690;
 g_Rooms = 0x801811AC;
 Random = 0x8018A168;
@@ -13,7 +14,9 @@ DestroyEntity = 0x8018D580;
 MoveEntity = 0x8018D934;
 FallEntity = 0x8018D964;
 SetStep = 0x8018E1C0;
+InitializeEntity = 0x8018E290;
 CollectSubweapon = 0x8018ED6C;
+EntityIntenseExplosion = 0x80192248;
 ClutLerp = 0x8019344C;
 EntitySoulStealOrb = 0x80194590;
 func_801951F0 = 0x801951F0;
