+++ src/st/cen/cen.h
@@ -91,5 +91,6 @@ extern u8 D_8019C770;
 
 extern u16 UNK_Update0[];
 extern u16 UNK_Invincibility0[];
+extern u16 g_InitializeEntityData0[];
 
 #endif
