+++ src/st/entity_intense_explosion.h
@@ -1,0 +1,35 @@
+void EntityIntenseExplosion(Entity* entity) {
+    u32 temp_v0;
+
+    if (entity->step == 0) {
+        InitializeEntity(g_InitializeEntityData0);
+        entity->palette = 0x8170;
+        entity->animSet = ANIMSET_DRA(5);
+        entity->animCurFrame = 1;
+        entity->blendMode = 0x30;
+
+        if (entity->params & 0xF0) {
+            entity->palette = 0x8195;
+            entity->blendMode = 0x10;
+        }
+
+        temp_v0 = entity->params & 0xFF00;
+        if (temp_v0 != 0) {
+            entity->zPriority = temp_v0 >> 8;
+        }
+
+        entity->zPriority += 8;
+        return;
+    }
+
+    entity->animFrameDuration++;
+    entity->posY.val -= FIX(0.25);
+
+    if (!(entity->animFrameDuration & 1)) {
+        entity->animCurFrame++;
+    }
+
+    if (entity->animFrameDuration >= 0x25) {
+        DestroyEntity(entity);
+    }
+}
