+++ src/st/np3/44DCC.c
@@ -694,7 +694,7 @@ void func_801C7E18(Entity* self) {
 
     switch (self->step) {
     case 0:
-        InitializeEntity(D_80180A54);
+        InitializeEntity(g_InitializeEntityData0);
         self->animSet = ANIMSET_DRA(2);
         self->velocityY = FIX(-5);
         self->palette = 0x8162;
