+++ src/st/np3/49F98.c
@@ -22,7 +22,7 @@ void EntityBloodSplatter(Entity* self) {
 
     switch (self->step) {
     case 0:
-        InitializeEntity(D_80180A54);
+        InitializeEntity(g_InitializeEntityData0);
         break;
 
     case 1:
@@ -449,7 +449,7 @@ void func_801CAE0C(Entity* self) { // BloodDrips
 
     switch (self->step) {
     case 0:
-        InitializeEntity(D_80180A54);
+        InitializeEntity(g_InitializeEntityData0);
         primIndex = g_api.AllocPrimitives(PRIM_LINE_G2, 1);
         if (primIndex != -1) {
             prim = &g_PrimBuf[primIndex];
