+++ src/st/nz0/47048.c
@@ -189,7 +189,7 @@ void func_801C7654(Entity* entity) {
 
     switch (entity->step) {
     case 0:
-        InitializeEntity(D_80180BE0);
+        InitializeEntity(g_InitializeEntityData0);
         entity->animSet = ANIMSET_DRA(2);
         entity->palette = 0x816D;
         entity->blendMode = 0x70;
