+++ src/st/nz0/nz0.h
@@ -45,7 +45,7 @@ extern u8 D_8003BE6F[];
 LayoutEntity* D_801808EC[];
 LayoutEntity* D_801809C0[];
 extern PfnEntityUpdate PfnEntityUpdates[];
-extern const u16 D_80180BE0[];
+extern const u16 g_InitializeEntityData0[];
 extern u16 D_80180BEC[];
 extern u16 D_80180BF8[]; // Init Elevator2
 extern u16 D_80180C1C[];
